Move vimrc to vim dir, update/organize bootstrap
Move the vimrc file into the vim dir, and update it so there's no internall hard coded references to the vimrc location. This *should* make it so we can put the general dotfiles directory anywhere, run bootstrap.sh, and everything just works.

Also updates the bootstrap.sh file with this change, and a few other tweaks:

* Set the `-n` flag on `ln` so that we don't follow symlinks (useful for when we're symlinking directories, like `sshrc.d`).
* Fold the `force` flag into the rest of the cli options instead of it hanging out by itself.
* Organize and comment `ln` commands for clarity