Git config template, inputrc fix, local bashrc fix
* The machine-local bashrc file is now loaded from ~/.local.bash, so
it's outside the git repo and at a conistent location (no uname lookup
required)
* Don't source inputrc, its not required, and apparently this messes up
some other things?
* Add a gitconfig.template file, which uses the [include] directive to
include the gitconfig.inc file. This template file is *COPIED* to the
homedir in bootstrap. This allows setting up the gitconfig with
generic configs from this repo, while allowing local config as well